// src/schemas/mst_mandatory_phic.ts
export interface mst_mandatory_phic {
    amount_start:                         number;
    amount_end :                          number;
    employee_contribution_value:          number;
    employer_contribution_value :         number;
    remarks :                             string;
    employee_contribution_percentage:     number;
    employer_contribution_percentage:     number;
  }

export interface mst_mandatory_phicID {
  id: number;
}